Improve UID de-id handling.
- Protect UIDs when mentioned in PS3.15 Table E1.1
- Additionally, since there are instance UIDs mentioned in PS3.6 but not PS3.15E1.1, protect all UIDs that are not official class UIDs.

The latter is done by checking against a global list of class UIDs and protecting if the UID does not match.
This errs on the side of privacy as we decide we'll protect a private Class UID even though it's a Class UID, since a privately generated UID can still likely be traced back.

For this mechanism, we create a script that creates the list of known class UIDs from the standard.