Update the apply method to accept either a directory or file list
The `organizeOptions` structure now has an option to specify
`inputFiles` instead of `inputDirectory` (mutually exclusive).

TODO: We need to provide a curation specification that works to
map only using information in the file name and the dicom header
since there is no directory path available when files are dropped
on the browser (only the file name component is available, not the
containing folder(s)).