feat: implement a first pass at rsync mode with upload to url
This implements the upload with mtime and hash info in metadata
and lays out the logic for different modes.
Optionally accepts a information about files and only performs
curation if the criteria do not match.  In basic mode, the size
and modification time are checked.  In"deep" mode the hash is also
checked.